+++ crml/cennzx-spot/src/lib.rs
@@ -23,16 +23,23 @@ mod tests;
 mod impls;
 mod types;
 pub use impls::{ExchangeAddressFor, ExchangeAddressGenerator};
-pub use types::FeeRate;
+pub use types::{FeeRate, U256};
 
 #[macro_use]
 extern crate srml_support as support;
 
 use generic_asset;
 use rstd::prelude::*;
-use runtime_primitives::traits::{As, Bounded, One, Zero};
-use support::{dispatch::Result, Dispatchable, Parameter, StorageDoubleMap, StorageMap, StorageValue};
+use runtime_primitives::traits::{Bounded, One, Zero};
+use support::{
+	dispatch::Result as DispatchResult, Dispatchable, Parameter, StorageDoubleMap, StorageMap, StorageValue,
+};
 use system::ensure_signed;
+
+// TODO: remove this temp fix for overflow issue when upstream fix ready
+#[macro_use]
+extern crate uint;
+use core::convert::TryInto;
 
 // (core_asset_id, asset_id)
 pub type ExchangeKey<T> = (
@@ -45,6 +52,11 @@ pub trait Trait: system::Trait + generic_asset::Trait {
 	type Event: From<Event<Self>> + Into<<Self as system::Trait>::Event>;
 	/// A function type to get an exchange address given the asset ID pair.
 	type ExchangeAddressGenerator: ExchangeAddressFor<Self::AssetId, Self::AccountId>;
+	// TODO: remove this temp fix for overflow issue when upstream fix ready
+	type BalanceToU128: From<<Self as generic_asset::Trait>::Balance> + Into<u128>;
+	type U128ToBalance: From<u128> + Into<<Self as generic_asset::Trait>::Balance>;
+
+	//	type AsBalance: From<<Self as generic_asset::Trait>::Balance> + Into<<Self as generic_asset::Trait>::Balance> + As<u128>;
 }
 
 decl_module! {
@@ -66,7 +78,7 @@ decl_module! {
 			#[compact] asset_bought: T::AssetId,
 			#[compact] buy_amount: T::Balance,
 			#[compact] max_paying_amount: T::Balance
-		) -> Result {
+		) -> DispatchResult {
 			let buyer = ensure_signed(origin)?;
 			let _ = Self::make_asset_swap_output(
 				&buyer,
@@ -95,7 +107,7 @@ decl_module! {
 			#[compact] asset_bought: T::AssetId,
 			#[compact] sell_amount: T::Balance,
 			#[compact] min_receive: T::Balance
-		) -> Result {
+		) -> DispatchResult {
 			let seller = ensure_signed(origin)?;
 			let _ = Self::make_asset_swap_input(
 				&seller,
@@ -182,7 +194,7 @@ decl_module! {
 			#[compact] liquidity_withdrawn: T::Balance,
 			#[compact] min_asset_withdraw: T::Balance,
 			#[compact] min_core_withdraw: T::Balance
-		) -> Result {
+		) -> DispatchResult {
 			let from_account = ensure_signed(origin)?;
 			ensure!(liquidity_withdrawn > Zero::zero(), "Amount of exchange asset to burn should exist");
 			ensure!(min_asset_withdraw > Zero::zero() && min_core_withdraw > Zero::zero(), "Assets withdrawn to be greater than zero");
@@ -213,7 +225,7 @@ decl_module! {
 		}
 
 		/// Set the spot exchange wide fee rate (root only)
-		pub fn set_fee_rate(new_fee_rate: FeeRate) -> Result {
+		pub fn set_fee_rate(new_fee_rate: FeeRate) -> DispatchResult {
 			<DefaultFeeRate<T>>::mutate(|fee_rate| *fee_rate = new_fee_rate);
 			Ok(())
 		}
@@ -229,11 +241,11 @@ decl_event!(
 		<T as generic_asset::Trait>::AssetId,
 		<T as generic_asset::Trait>::Balance
 	{
-		// Provider, core asset amount, trade asset id, trade asset amount
+		/// Provider, core asset amount, trade asset id, trade asset amount
 		AddLiquidity(AccountId, Balance, AssetId, Balance),
-		// Provider, core asset amount, trade asset id, trade asset amount
+		/// Provider, core asset amount, trade asset id, trade asset amount
 		RemoveLiquidity(AccountId, Balance, AssetId, Balance),
-	    // AssetSold, AssetBought, Buyer, SoldAmount, BoughtAmount
+	    /// AssetSold, AssetBought, Buyer, SoldAmount, BoughtAmount
 		AssetPurchase(AssetId, AssetId, AccountId, Balance, Balance),
 	}
 );
@@ -625,12 +637,7 @@ impl<T: Trait> Module<T> {
 
 		let core_reserve = <generic_asset::Module<T>>::free_balance(&core_asset_id, &exchange_address);
 
-		Ok(Self::get_output_price(
-			buy_amount,
-			core_reserve,
-			asset_reserve,
-			fee_rate,
-		))
+		Self::get_output_price(buy_amount, core_reserve, asset_reserve, fee_rate)
 	}
 
 	/// `asset_id` - Trade asset
@@ -648,12 +655,7 @@ impl<T: Trait> Module<T> {
 
 		let asset_reserve = <generic_asset::Module<T>>::free_balance(asset_id, &exchange_address);
 		let core_reserve = <generic_asset::Module<T>>::free_balance(&core_asset_id, &exchange_address);
-		Ok(Self::get_input_price(
-			sell_amount,
-			asset_reserve,
-			core_reserve,
-			fee_rate,
-		))
+		Self::get_input_price(sell_amount, asset_reserve, core_reserve, fee_rate)
 	}
 
 	fn get_output_price(
@@ -661,43 +663,56 @@ impl<T: Trait> Module<T> {
 		input_reserve: T::Balance,
 		output_reserve: T::Balance,
 		fee_rate: FeeRate,
-	) -> T::Balance {
+	) -> rstd::result::Result<T::Balance, &'static str> {
 		if input_reserve.is_zero() || output_reserve.is_zero() {
-			return Zero::zero();
+			return Err("Pool is empty");
 		}
 
 		// Special case, in theory price should progress towards infinity
 		if output_amount >= output_reserve {
-			return T::Balance::max_value();
+			return Ok(T::Balance::max_value());
 		}
 
-		let numerator: T::Balance = input_reserve * output_amount;
-		let denominator = output_reserve - output_amount;
-		let output: T::Balance = numerator / denominator + One::one();
+		let amount = U256::from(T::BalanceToU128::from(output_amount).into());
+		let input_reserve = U256::from(T::BalanceToU128::from(input_reserve).into());
+		let denominator = U256::from(T::BalanceToU128::from(output_reserve - output_amount).into());
 
-		(FeeRate::one() + fee_rate) * output
+		let res: u128 = (input_reserve * amount / denominator).try_into().map_err(|_| "Overflow error")?;
+
+		let price = T::U128ToBalance::from(res).into();
+		let price_plus_one = T::BalanceToU128::from(price + One::one());
+		let output = FeeRate::safe_mul(FeeRate::one() + fee_rate, price_plus_one)?;
+		Ok(T::U128ToBalance::from(output).into())
 	}
+
 
 	fn get_input_price(
 		input_amount: T::Balance,
 		input_reserve: T::Balance,
 		output_reserve: T::Balance,
 		fee_rate: FeeRate,
-	) -> T::Balance {
+	) -> rstd::result::Result<T::Balance, &'static str> {
 		if input_reserve.is_zero() || output_reserve.is_zero() {
-			return Zero::zero();
+			return Err("Pool is empty");
 		}
-		let div_rate = FeeRate::one() + fee_rate;
-		// This operation rounds away necessary decimal points. In order to-
-		// counteract this, we scale the input amount
-		let input_amount_less_fee_scaled = FeeRate::div(
-			input_amount * T::Balance::sa(1_000_000), // scale up
-			div_rate,
-		);
-		let numerator: T::Balance = input_amount_less_fee_scaled * output_reserve;
-		let denominator: T::Balance = FeeRate::div(input_amount, div_rate) + input_reserve;
 
-		numerator / denominator / T::Balance::sa(1_000_000) // undo scaling
+		let div_rate = FeeRate::one() + fee_rate;
+
+		let lhs = T::BalanceToU128::from(input_amount);
+
+		let input_amount_less_fee_scaled = FeeRate::safe_div(
+			lhs,
+			div_rate,
+		)?;
+		let input_reserve: u128 = T::BalanceToU128::from(input_reserve).into();
+		let output_reserve = U256::from(T::BalanceToU128::from(output_reserve).into());
+		let input_amount = U256::from(input_amount_less_fee_scaled);
+
+		let denominator: u128 = (input_amount + U256::from(input_reserve)).try_into().map_err(|_| "Overflow error")?;
+
+		let res: u128 = (output_reserve * input_amount / denominator).try_into().map_err(|_| "Overflow error")?;
+
+		Ok(T::U128ToBalance::from(res).into())
 	}
 
 	/// `asset_id` - Trade asset
@@ -722,12 +737,7 @@ impl<T: Trait> Module<T> {
 
 		let trade_asset_reserve = <generic_asset::Module<T>>::free_balance(&asset_id, &exchange_address);
 
-		Ok(Self::get_output_price(
-			buy_amount,
-			trade_asset_reserve,
-			core_asset_reserve,
-			fee_rate,
-		))
+		Self::get_output_price(buy_amount, trade_asset_reserve, core_asset_reserve, fee_rate)
 	}
 
 	/// Returns the amount of trade asset to pay for `sell_amount` of core sold.
@@ -747,7 +757,7 @@ impl<T: Trait> Module<T> {
 		let core_asset_reserve = <generic_asset::Module<T>>::free_balance(&core_asset_id, &exchange_address);
 		let trade_asset_reserve = <generic_asset::Module<T>>::free_balance(asset_id, &exchange_address);
 
-		let output_amount = Self::get_input_price(sell_amount, core_asset_reserve, trade_asset_reserve, fee_rate);
+		let output_amount = Self::get_input_price(sell_amount, core_asset_reserve, trade_asset_reserve, fee_rate)?;
 
 		ensure!(
 			trade_asset_reserve > output_amount,
@@ -774,7 +784,7 @@ impl<T: Trait> Module<T> {
 		buy_amount: T::Balance,
 		max_paying_amount: T::Balance,
 		fee_rate: FeeRate,
-	) -> Result {
+	) -> DispatchResult {
 		let core_asset = Self::core_asset_id();
 		ensure!(asset_sold != asset_bought, "Asset to swap should not be equal");
 		if *asset_sold == core_asset {
@@ -821,7 +831,7 @@ impl<T: Trait> Module<T> {
 		sell_amount: T::Balance,
 		min_receive: T::Balance,
 		fee_rate: FeeRate,
-	) -> Result {
+	) -> DispatchResult {
 		let core_asset = Self::core_asset_id();
 		ensure!(asset_sold != asset_bought, "Asset to swap should not be equal");
 		if *asset_sold == core_asset {